whisper-modeling/requirements.txt
loralib==0.1.2
numpy==1.24.4
pandas==2.0.3
pytorch_lightning==2.2.4
PyYAML==6.0.2
scikit_learn==1.3.2
torch==2.3.0
torchaudio==2.3.0
tqdm==4.66.4
transformers==4.30.2
tensorboard